refactor: Product 리팩토링
