+++ src/main/java/kitchenpos/product/tobe/application/dto/ChangeProductPriceRequestDto.java
@@ -1,0 +1,15 @@
+package kitchenpos.product.tobe.application.dto;
+
+import java.math.BigDecimal;
+
+public class ChangeProductPriceRequestDto {
+    private final BigDecimal price;
+
+    public ChangeProductPriceRequestDto(BigDecimal price) {
+        this.price = price;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
+    }
+}
