+++ src/main/java/kitchenpos/product/tobe/infra/JpaProductRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.product.infra;
+package kitchenpos.product.tobe.infra;
 
 import java.util.UUID;
-import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductRepository;
+import kitchenpos.product.tobe.domain.Product;
+import kitchenpos.product.tobe.domain.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaProductRepository extends ProductRepository, JpaRepository<Product, UUID> {
